// src/commons/connector/yaci/types/block-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Yaci Store API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { Vrf } from "./vrf";
/**
 *
 *
 * @export
 * @interface BlockDto
 */
export interface BlockDto {
  /**
   * @type {number}
   * @memberof BlockDto
   */
  time?: number;

  /**
   * @type {number}
   * @memberof BlockDto
   */
  height?: number;

  /**
   * @type {number}
   * @memberof BlockDto
   */
  number?: number;

  /**
   * @type {string}
   * @memberof BlockDto
   */
  hash?: string;

  /**
   * @type {number}
   * @memberof BlockDto
   */
  slot?: number;

  /**
   * @type {number}
   * @memberof BlockDto
   */
  epoch?: number;

  /**
   * @type {number}
   * @memberof BlockDto
   */
  era?: number;

  /**
   * @type {number}
   * @memberof BlockDto
   */
  epochSlot?: number;

  /**
   * @type {string}
   * @memberof BlockDto
   */
  slotLeader?: string;

  /**
   * @type {number}
   * @memberof BlockDto
   */
  size?: number;

  /**
   * @type {number}
   * @memberof BlockDto
   */
  txCount?: number;

  /**
   * @type {number}
   * @memberof BlockDto
   */
  output?: number;

  /**
   * @type {number}
   * @memberof BlockDto
   */
  fees?: number;

  /**
   * @type {string}
   * @memberof BlockDto
   */
  blockVrf?: string;

  /**
   * @type {string}
   * @memberof BlockDto
   */
  opCert?: string;

  /**
   * @type {number}
   * @memberof BlockDto
   */
  opCertCounter?: number;

  /**
   * @type {number}
   * @memberof BlockDto
   */
  opCertKesPeriod?: number;

  /**
   * @type {string}
   * @memberof BlockDto
   */
  opCertSigma?: string;

  /**
   * @type {string}
   * @memberof BlockDto
   */
  previousBlock?: string;

  /**
   * @type {string}
   * @memberof BlockDto
   */
  issuerVkey?: string;

  /**
   * @type {Vrf}
   * @memberof BlockDto
   */
  nonceVrf?: Vrf;

  /**
   * @type {Vrf}
   * @memberof BlockDto
   */
  leaderVrf?: Vrf;

  /**
   * @type {Vrf}
   * @memberof BlockDto
   */
  vrfResult?: Vrf;

  /**
   * @type {string}
   * @memberof BlockDto
   */
  blockBodyHash?: string;

  /**
   * @type {string}
   * @memberof BlockDto
   */
  protocolVersion?: string;
}