// src/commons/connector/yaci/types/blocks-page.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Yaci Store API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { BlockSummary } from "./block-summary";
/**
 *
 *
 * @export
 * @interface BlocksPage
 */
export interface BlocksPage {
  /**
   * @type {number}
   * @memberof BlocksPage
   */
  total?: number;

  /**
   * @type {number}
   * @memberof BlocksPage
   */
  totalPages?: number;

  /**
   * @type {Array<BlockSummary>}
   * @memberof BlocksPage
   */
  blocks?: Array<BlockSummary>;
}