// src/commons/connector/yaci/types/transaction-summary.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Yaci Store API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/**
 *
 *
 * @export
 * @interface TransactionSummary
 */
export interface TransactionSummary {
  /**
   * @type {string}
   * @memberof TransactionSummary
   */
  txHash?: string;

  /**
   * @type {number}
   * @memberof TransactionSummary
   */
  blockNumber?: number;

  /**
   * @type {number}
   * @memberof TransactionSummary
   */
  slot?: number;

  /**
   * @type {Array<string>}
   * @memberof TransactionSummary
   */
  outputAddresses?: Array<string>;

  /**
   * @type {number}
   * @memberof TransactionSummary
   */
  totalOutput?: number;

  /**
   * @type {number}
   * @memberof TransactionSummary
   */
  fee?: number;
}