// src/commons/connector/yaci/types/epoch.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Yaci Store API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/**
 *
 *
 * @export
 * @interface Epoch
 */
export interface Epoch {
  /**
   * @type {number}
   * @memberof Epoch
   */
  number?: number;

  /**
   * @type {number}
   * @memberof Epoch
   */
  blockCount?: number;

  /**
   * @type {number}
   * @memberof Epoch
   */
  transactionCount?: number;

  /**
   * @type {number}
   * @memberof Epoch
   */
  totalOutput?: number;

  /**
   * @type {number}
   * @memberof Epoch
   */
  totalFees?: number;

  /**
   * @type {number}
   * @memberof Epoch
   */
  startTime?: number;

  /**
   * @type {number}
   * @memberof Epoch
   */
  endTime?: number;

  /**
   * @type {number}
   * @memberof Epoch
   */
  maxSlot?: number;
}