// src/commons/connector/yaci/types/transaction-page.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Yaci Store API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { TransactionSummary } from "./transaction-summary";
/**
 *
 *
 * @export
 * @interface TransactionPage
 */
export interface TransactionPage {
  /**
   * @type {number}
   * @memberof TransactionPage
   */
  total?: number;

  /**
   * @type {number}
   * @memberof TransactionPage
   */
  totalPages?: number;

  /**
   * @type {Array<TransactionSummary>}
   * @memberof TransactionPage
   */
  transactionSummaries?: Array<TransactionSummary>;
}